Reorganization of a few classes and files.
