+++ src/main/kotlin/network/data/messages/PacketBuilder.kt
@@ -1,4 +1,4 @@
-package network.data.communication
+package network.data.messages
 
 import Configuration
 import network.data.Endpoint
